cast where appropriate; might fix a bad bug in PaddedEphemerallyEncryptedBucket
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@19298 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/node/UptimeEstimator.java
@@ -101,7 +101,7 @@ public class UptimeEstimator implements Runnable {
 			if(slot == wasOnline.length) slot = 0;
 		}
 		long now = System.currentTimeMillis();
-		if(logFile.length() > wasOnline.length*4) {
+		if(logFile.length() > (long) wasOnline.length*4) {
 			prevFile.delete();
 			logFile.renameTo(prevFile);
 		}
